requirements.txt
aioredis==1.3.1
appdirs==1.4.4
apturl==0.5.2
asgiref==3.3.1
asttokens==2.0.8
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.6.0
azure-core==1.20.1
azure-storage-blob==12.9.0
backcall==0.2.0
bcrypt==3.1.7
beautifulsoup4==4.11.1
black==21.6b0
blinker==1.4
Brlapi==0.7.0
Brotli==1.0.9
CacheControl==0.12.10
cachetools==4.2.4
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
command-not-found==0.3
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cowsay==4.0
cryptography==3.4.7
cssselect2==0.4.1
cupshelpers==1.0
daphne==3.0.2
dbus-python==1.2.16
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==3.1.7
django-cors-headers==3.7.0
django-crontab==0.7.1
django-debug-toolbar==3.2.1
django-filter==2.4.0
django-storages==1.12.3
djangorestframework==3.12.2
drf-yasg==1.20.0
duplicity==0.8.12.0
entrypoints==0.3
et-xmlfile==1.1.0
executing==1.1.0
fasteners==0.14.1
filelock==3.0.12
firebase-admin==5.1.0
fonttools==4.25.2
future==0.18.2
gcloud==0.18.3
google-api-core==2.2.2
google-api-python-client==2.31.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
google-cloud-core==2.2.1
google-cloud-firestore==2.3.4
google-cloud-storage==1.43.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
grpcio==1.42.0
grpcio-status==1.42.0
gunicorn==20.1.0
hiredis==2.0.0
html5lib==1.1
httplib2==0.19.1
hyperlink==21.0.0
idna==2.10
importlib-metadata==4.6.1
incremental==21.3.0
inflection==0.5.1
iniconfig==1.1.1
ipython==8.5.0
isodate==0.6.0
isort==5.9.1
itypes==1.2.0
jedi==0.18.1
Jinja2==2.11.3
jws==0.1.3
kafka-python==2.0.2
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
lxml==4.9.1
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib-inline==0.1.6
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.2
msrest==0.6.21
mypy-extensions==0.4.3
netifaces==0.10.4
numpy==1.21.1
oauth2client==4.1.3
oauthlib==3.1.1
odfpy==1.4.1
olefile==0.46
openpyxl==3.0.7
packaging==20.9
pandas==1.3.1
paramiko==2.6.0
parso==0.8.3
pathspec==0.8.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==8.2.0
pipenv==11.9.0
pluggy==1.0.0
prompt-toolkit==3.0.31
proto-plus==1.19.8
protobuf==3.19.1
psycopg2-binary==2.8.6
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.16.2
pycodestyle==2.8.0
pycparser==2.20
pycryptodome==3.11.0
pycups==1.9.73
pydyf==0.0.3
pygame==2.1.2
Pygments==2.13.0
PyGObject==3.36.0
PyJWT==2.0.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyphen==0.11.0
Pyrebase4==4.5.0
pyRFC3339==1.1
pytest==7.1.2
python-apt==2.0.0+ubuntu0.20.4.8
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-dotenv==0.18.0
python-http-client==3.3.2
python-jwt==2.0.1
pytz==2021.1
pyxdg==0.26
PyYAML==5.4.1
regex==2021.7.6
reportlab==3.5.34
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
requests-unixsocket==0.2.0
rsa==4.7.2
ruamel.yaml==0.16.13
ruamel.yaml.clib==0.2.2
SecretStorage==2.3.1
sendgrid==6.8.1
service-identity==21.1.0
simplejson==3.16.0
six==1.16.0
smpplib==2.1.0
soupsieve==2.3.2.post1
sqlparse==0.4.1
stack-data==0.5.1
starkbank-ecdsa==1.1.1
systemd-python==234
tinycss2==1.1.0
toml==0.10.2
tomli==2.0.1
traitlets==5.4.0
Twisted==21.2.0
txaio==21.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.0
ua-parser==0.10.0
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.26.3
usb-creator==0.3.7
user-agents==2.2.0
virtualenv==20.0.17
virtualenv-clone==0.3.0
wadllib==1.3.3
wcwidth==0.2.5
WeasyPrint==52.5
webencodings==0.5.1
xkit==0.0.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.5.0
zope.interface==5.4.0
zopfli==0.1.8
